# Frontend Docker build improvements