--- a/frontend/Dockerfile
+++ b/frontend/Dockerfile
@@ -3,14 +3,6 @@ ARG RWP_BASE_URL=https://cdn.jsdelivr.net/npm/replaywebpage/
 
 FROM docker.io/library/node:16 as build_deps
 
-ARG GIT_COMMIT_HASH
-ARG GIT_BRANCH_NAME
-ARG VERSION
-
-ENV GIT_COMMIT_HASH=${GIT_COMMIT_HASH} \
-    GIT_BRANCH_NAME=${GIT_BRANCH_NAME} \
-    VERSION=${VERSION}
-
 WORKDIR /app
 COPY yarn.lock package.json ./
 RUN yarn --frozen-lockfile --network-timeout 1000000 && \
@@ -28,6 +20,17 @@ COPY lit-localize.json \
     ./
 
 COPY src ./src/
+
+# Build variables used to show current app version
+# in the UI. Note that this will invalidate all
+# subsequent RUN steps.
+ARG GIT_COMMIT_HASH
+ARG GIT_BRANCH_NAME
+ARG VERSION
+
+ENV GIT_COMMIT_HASH=${GIT_COMMIT_HASH} \
+    GIT_BRANCH_NAME=${GIT_BRANCH_NAME} \
+    VERSION=${VERSION}
 
 RUN yarn build