Improves caching performance and decreases size of build image:
- Does not bundle playwright or test runners, only build dependencies
- Moves `yarn install` to separate `build_deps` stage and removes yarn cache from the build image in favor of Docker layer caching
- Moves version ARGS & ENV to right before `yarn build` to prevent cache invalidation on yarn install step
- Uses [`COPY --link`](https://docs.docker.com/engine/reference/builder/#copy---link)

Testing locally with `docker buildx build -t local/browsertrix-cloud-frontend:latest .`:
Before: >400s every build (I stopped it at 5 minutes locally)
After: <100s first build

Testing with gh action:
Before: [5m 23s](https://github.com/webrecorder/browsertrix-cloud/actions/runs/4602484842/jobs/8131491015#step:7:256)
After: [1m 53s](https://github.com/webrecorder/browsertrix-cloud/actions/runs/4622847011/jobs/8176035787#step:7:279)

Down the line we may want to upgrade to yarn >=2 which has better docker support.