@ikreymer 2023-04-06
Nice this is great!
A further optimization might be to just put all of the build time vars VERSION, GIT_COMMIT_HASH, GIT_BRANCH_NAME, and RWP_BASE_URL into a json file that is just loaded on the frontend? that way doesn't have to be part of the build at all..
(can be a separate change later)